+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 1.0b1 (unreleased)
 ------------------
 
+- Icalendar export of attendees almost according to the RFC5545 standard. At
+  the moment, we do not distinguish between CN and CAL-ADDRESS in Plone, so we
+  just put the attendee value to the CN and CAL-ADDRESS parameter. Fixes #24.
+  [thet]
+
 - Support microseconds for DateTime conversions. For recurrence rules,
   timezones are not supported due to a python-datetime limitation.
   [thet]